pyproject.toml
[project]
name = "sheffybot"
version = "0.1.0"
description = "sheffybot"
readme = "README.md"
requires-python = ">=3.10,<4.0"
authors = [{ name = "liangzimiao", email = "2968274386@qq.com" }]
license = { text = "Apache-2.0" }
dependencies = [
    "nonebot2[fastapi,httpx,websockets]<3.0.0,>=2.0.1",
    "nonebot-adapter-qq>=1.1.0",
    "nonebot-plugin-alconna<1.0.0,>=0.31.0",
    "nonebot-plugin-apscheduler<1.0.0,>=0.3.0",
]


[project.optional-dependencies]
PCR = [
    "requests<3.0.0,>=2.31.0",
    "pillow>=10.1.0",
    "pytz>=2023.3.post1",
    "fuzzywuzzy>=0.18.0",
    "zhconv>=1.4.3",
]


[tool.nonebot]
adapters = [{ name = "QQ", module_name = "nonebot.adapters.qq" }]
plugins = ["nonebot_plugin_alconna", "nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]


[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
